requirements/base.txt
asgiref==3.7.2
contourpy==1.1.1
cycler==0.11.0
Django==4.2.5
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-js-asset==2.1.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
